moved carousel.js
