+++ app/javascript/components/index.js
@@ -1,0 +1,3 @@
+import initCarousel from "./carousel";
+
+initCarousel();
