+++ app/javascript/packs/application.js
@@ -14,3 +14,4 @@ ActiveStorage.start()
 
 import "controllers"
 import "bootstrap"
+import "components"
